recipes/natural_language_processing/graph-rag/app/requirements.txt
ruamel.yaml==0.18.10
streamlit==1.42.0
graphrag==1.2.0
PyMuPDF==1.25.3